% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gmtr.R
\name{calc_gmtr}
\alias{calc_gmtr}
\title{Calculate GMTR function}
\usage{
calc_gmtr(
  data,
  var_expr,
  group_expr,
  method_ci = "t.test",
  decimals = 1,
  groups_order,
  nboot = 1000,
  ci_sep = "‒",
  dec_sep = ".",
  base = exp(1)
)
}
\arguments{
\item{data}{A tibble with the data to be used.}

\item{var_expr}{A name with the name of the variable in data.}

\item{group_expr}{A name with the name of the group variable in data.}

\item{method_ci}{A character with method used to calculate CI. Use 't.test' or 'bootstrap'.}

\item{decimals}{A number with the number of decimals points to present the results.}

\item{groups_order}{A character array with the names of the groups in order to be used in GMTR calculation. GMTR = First group / Second Group.}

\item{nboot}{A number with the number of replications to be applied if bootstrap is selected.}

\item{ci_sep}{A character with a separator symbol to be used in CI.}

\item{dec_sep}{A character to specify the decimal separator to be used.}

\item{base}{A number with the number to be used as base in the antilogarithm calculation.}
}
\value{
A tibble with the GMTR results.
}
\description{
Calculate GMTR function
}
\examples{
data <- tibble(titer = c(rnorm(100),rbeta(100, 0.2, 0.2)), group = c(rep(1, 100), rep(2, 100)))
calc_gmtr(data = data, var_expr = titer, group_expr = group, groups_order = c('2', '1'), method_ci = 't.test', decimals = 2, ci_sep = '\u2012', dec_sep = ".", base = exp(1))
}
