% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_medqrt.R
\name{calc_medqrt}
\alias{calc_medqrt}
\title{Calculate Median and Q1 and Q3 quantiles function}
\usage{
calc_medqrt(
  data,
  var_expr,
  group_expr,
  decimals = 1,
  groups_order = NA_character_,
  dec_sep = ".",
  type = 7,
  sep = "‒"
)
}
\arguments{
\item{data}{A tibble with the data to be used.}

\item{var_expr}{A name with the name of the variable in data.}

\item{group_expr}{A name with the name of the group variable in data.}

\item{decimals}{A number with the number of decimals points to present the results.}

\item{groups_order}{A character array with the names of the groups in order to be presented in the results.}

\item{dec_sep}{A character to specify the decimal separator to be used.}

\item{type}{A number between 1 and 9 selecting one of the nine quantile algorithms detailed in quantile function to be used.}

\item{sep}{A character with a separator symbol}
}
\value{
A tibble with the Median and Q1 and Q3 quantiles results by group.
}
\description{
Calculate Median and Q1 and Q3 quantiles function
}
\examples{
data <- tibble(titer = c(rnorm(100),rbeta(100, 0.2, 0.2)), group = c(rep(1, 100), rep(2, 100)))
calc_medqrt(data, titer, group,  dec_sep = '.', decimals = 2, groups_order = c('2', '1'), type = 7, sep = ';')
}
