% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller_gmtr.R
\name{caller_gmtr}
\alias{caller_gmtr}
\title{Calculate GMTR}
\usage{
caller_gmtr(
  data,
  titer_expr,
  group_expr,
  groups_order,
  force_ci = NA_character_,
  decimals = 1,
  nboot = 1000
)
}
\arguments{
\item{data}{A tibble with the data to be used.}

\item{titer_expr}{A name with the name of the titer variable in data.}

\item{group_expr}{A name with the name of the group variable in data. Number of groups must be two.}

\item{groups_order}{A character array with the names of the groups in order to be used in GMTR calculation. GMTR = First group / Second Group.}

\item{force_ci}{A character to force method used to calculate CI. Use 't.test' or 'bootstrap'.}

\item{decimals}{A number with the number of decimals points to present the results.}

\item{nboot}{The number of replications to be applied if bootstrap is selected.}
}
\value{
A tibble with the GMTR results.
}
\description{
Calculate GMTR
}
\examples{
data <- tibble(titer = c(rnorm(100),rbeta(100, 0.2, 0.2)), group = c(rep(1, 100), rep(2, 100)))
caller_gmtr(data, titer, group, groups_order = c('2', '1'), decimals = 2)
}
