% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller_medq1q3.R
\name{caller_medq1q3}
\alias{caller_medq1q3}
\title{Calculate Median and Q1 and Q3 quantiles}
\usage{
caller_medq1q3(
  data,
  var_expr,
  group_expr,
  decimals = 1,
  groups_order = NA_character_
)
}
\arguments{
\item{data}{A tibble with the data to be used.}

\item{var_expr}{A name with the name of the variable in data.}

\item{group_expr}{A name with the name of the group variable in data.}

\item{decimals}{A number with the number of decimals points to present the results.}

\item{groups_order}{A character array with the names of the groups in order to be presented in the results.}
}
\value{
A tibble with the Median and Q1 and Q3 quantiles results by group.
}
\description{
Calculate Median and Q1 and Q3 quantiles
}
\examples{
data <- tibble(titer = c(rnorm(100),rbeta(100, 0.2, 0.2)), group = c(rep(1, 100), rep(2, 100)))
caller_medq1q3(data, titer, group, decimals = 2, groups_order = c('2', '1'))
}
