% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round1.R
\name{round1}
\alias{round1}
\title{Round function to follow IB Stats POP.}
\usage{
round1(x, digits = 1)
}
\arguments{
\item{x}{A number to round.}

\item{digits}{A number with number of digits to round.}
}
\value{
A rounded number.
}
\description{
Round function to follow IB Stats POP.
}
\examples{
x <- round1(0.456, 1)
}
