% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round1_str.R
\name{round1_str}
\alias{round1_str}
\title{Round1 function with conversion to string. 'This function is deprecated. Please, use `roundmath_str` function instead.'}
\usage{
round1_str(x, decimals = 1)
}
\arguments{
\item{x}{A number or array to be rounded.}

\item{decimals}{A number with de number of decimals points to round.}
}
\value{
Rounded number.
}
\description{
Round1 function with conversion to string. 'This function is deprecated. Please, use `roundmath_str` function instead.'
}
\examples{
data <- tibble(titer = c(rnorm(100),rbeta(100, 0.2, 0.2)), group = c(rep(1, 100), rep(2, 100)))
caller_gmtr(titer, group, groups_order = c('2', '1'), data, decimals = 2)
}
