% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundmath.R
\name{roundmath}
\alias{roundmath}
\title{Mathematical rounding function}
\usage{
roundmath(x, decimals = 1)
}
\arguments{
\item{x}{A number or array to be rounded.}

\item{decimals}{A number with the number of decimals points to round.}
}
\value{
Rounded number.
}
\description{
Mathematical rounding function
}
\examples{
x <- roundmath(x = 0.456, decimals = 1)
}
